+++ src/main/java/org/recast/Recast/Include/rcCompactSpan.java
@@ -2,7 +2,7 @@ package org.recast.Recast.Include;
 
 /// Represents a span of unobstructed space within a compact heightfield.
 public class rcCompactSpan {
-    public short y;			///< The lower extent of the span. (Measured from the heightfield's base.)
+    public int y;			///< The lower extent of the span. (Measured from the heightfield's base.)
     public int reg;			///< The id of the region the span belongs to. (Or zero if not in a region.)
     private int con/* = 24*/;		///< Packed neighbor connection data.
 //    private int con/* = 24*/;		///< Packed neighbor connection data.
