+++ src/main/java/org/recast/Recast/Include/rcEdge.java
@@ -2,6 +2,6 @@ package org.recast.Recast.Include;
 
 public class rcEdge {
     public int vert[] = new int[2];
-    public short polyEdge[] = new short[2];
-    public short poly[] = new short[2];
+    public int polyEdge[] = new int[2];
+    public int poly[] = new int[2];
 }
