+++ src/main/java/org/recast/Recast/Include/rcSweepSpan.java
@@ -3,6 +3,6 @@ package org.recast.Recast.Include;
 public class rcSweepSpan {
     public int rid;	// row id
     public int id;	// region id
-    public short ns;	// number samples
+    public int ns;	// number samples
     public int nei;	// neighbour id
 }
